catalog_service/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy[asyncio]
asyncpg
alembic
python-dotenv